// services/src/types.ts
export interface SearchSlice {
  searchQuery: string;
  isSearching: boolean;
  searchResults: string;
  showSearchResults: boolean;
  setSearchQuery: (query: string) => void;
  executeSearch: () => Promise<void>;
  closeSearchResults: () => void;
}
// ...
export interface AppState extends
    UISlice,
    ChatSlice,
    VoiceSlice,
    LiveSlice,
    KernelSlice,
    ForgeSlice,
    StudioSlice,
    CodeSlice,
    ImageSlice,
    TaskSlice,
    ConstitutionSlice,
    ResumeSlice,
    NotificationSlice,
    SearchSlice,
    AwesomeResourceSlice,
    DiarySlice,
    SnapshotSlice {}

export type AppSlice<T> = StateCreator<AppState, [['zustand/persist', unknown]], [], T>;